# db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.28.0

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Affiliate struct {
	ID              pgtype.UUID `json:"id"`
	Name            string      `json:"name"`
	MasterAffiliate pgtype.UUID `json:"master_affiliate"`
	Balance         float64     `json:"balance"`
}

type Commission struct {
	ID          pgtype.UUID `json:"id"`
	OrderID     pgtype.UUID `json:"order_id"`
	AffiliateID pgtype.UUID `json:"affiliate_id"`
	Amount      float64     `json:"amount"`
}

type Product struct {
	ID       pgtype.UUID `json:"id"`
	Name     string      `json:"name"`
	Quantity int32       `json:"quantity"`
	Price    float64     `json:"price"`
}

type User struct {
	ID          pgtype.UUID `json:"id"`
	Username    string      `json:"username"`
	Password    string      `json:"password"`
	Balance     float64     `json:"balance"`
	AffiliateID pgtype.UUID `json:"affiliate_id"`
}